+++ hypervisor/include/debug/vuart.h
@@ -75,6 +75,6 @@ struct acrn_vuart *vuart_console_active(void);
 void vuart_console_tx_chars(struct acrn_vuart *vu);
 void vuart_console_rx_chars(struct acrn_vuart *vu);
 
-bool hv_used_dbg_intx(uint8_t intx_pin);
+bool hv_used_dbg_intx(uint32_t intx_pin);
 void vuart_set_property(const char *vuart_info);
 #endif /* VUART_H */
